// @music/data-etl/src/generated/artist.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Artist {
  images?: {
    image?:
      | {
          "@type"?: string;
          "@uri"?: string;
          "@uri150"?: string;
          "@width"?: string;
          "@height"?: string;
          [k: string]: unknown;
        }
      | {
          "@type"?: string;
          "@uri"?: string;
          "@uri150"?: string;
          "@width"?: string;
          "@height"?: string;
          [k: string]: unknown;
        }[];
    [k: string]: unknown;
  };
  id?: string;
  name?: string | null;
  realname?: string | null;
  profile?: string | null;
  data_quality?: string;
  urls?: {
    url?: (string | null) | (string | null)[];
    [k: string]: unknown;
  };
  namevariations?: {
    name?: string | string[];
    [k: string]: unknown;
  };
  aliases?: null | {
    name?:
      | {
          "@id"?: string;
          "#text"?: string;
          [k: string]: unknown;
        }
      | {
          "@id"?: string;
          "#text"?: string;
          [k: string]: unknown;
        }[];
    [k: string]: unknown;
  };
  members?: {
    id?: string | string[];
    name?:
      | {
          "@id"?: string;
          "#text"?: string;
          [k: string]: unknown;
        }
      | {
          "@id"?: string;
          "#text"?: string;
          [k: string]: unknown;
        }[];
    [k: string]: unknown;
  };
  groups?: null | {
    name?:
      | {
          "@id"?: string;
          "#text"?: string;
          [k: string]: unknown;
        }
      | {
          "@id"?: string;
          "#text"?: string;
          [k: string]: unknown;
        }[];
    [k: string]: unknown;
  };
  [k: string]: unknown;
}